requirements.txt
# STT 음성 텍스트 변환 및 문제 생성 프로젝트 의존성

# 음성 처리 관련 (STT)
openai-whisper>=20250625
torch>=2.8.0
torchaudio>=2.8.0

# LLM API 통신 관련
ollama>=0.5.3
openai>=1.102.0
google-generativeai>=0.8.5

# 설정 및 환경변수 관리
python-dotenv>=1.1.1
PyYAML>=6.0.1

# YouTube 다운로드 관련
yt-dlp>=2025.8.22

# UI 및 터미널 출력
rich>=14.1.0

# 기본 유틸리티
numpy>=2.2.6
requests>=2.32.5

# 오디오/비디오 처리 (FFmpeg 의존성)
ffmpeg-python>=0.2.0

# 선택사항 - GPU 가속 (NVIDIA GPU 사용시)
# torch>=2.8.0+cu118
# torchaudio>=2.8.0+cu118

# 개발 도구 (선택사항)
# pytest>=8.0.0
# black>=24.0.0
# flake8>=7.0.0
# mypy>=1.8.0
